Added test for dependent script APT packages and solved errors with dependencies in some scripts.

+++ src/unix/debian/install/circleci/main.sh
@@ -28,6 +28,10 @@ function printPrependedStdout() {
   printf "%s" "$_PREPEND_STDOUT_STRING"
 }
 
+if [[ "$(sudo dpkg -s curl 2> /dev/null | grep Status)" != "Status: install ok installed" ]]; then
+  sudo apt-get install -y -qqq curl > /dev/null || exit $?
+fi;
+
 _CIRCLECI_BINARY_PATH=$(which circleci)
 _CIRCLECI_LATEST_RELEASE_VERSION=$(
   curl -sLI -o /dev/null -w '%{url_effective}' \
